Resolve issue where a link in the dialog was causing a full page reload.
The dialog was stopping event propagation which prevented kit from
using pushstate to handle the link click -- the browser's default
behavior was allowed to happen.

The stop propagation was in place to prevent clicks within the dialog
content from being interpretted as clicking outside the dialog.

It has been tweaked so the only way to click the dialog is by clicking
its ::backdrop, which really is clicking outside the dialog.